Refactored databse helpers to be generic and made wrappers for the movie fetching functions.
