+++ src/app/api/requests/testconnection.ts
@@ -1,10 +1,10 @@
 "use server";
 import db from "@/lib/mongodb";
 
-export async function testDatabaseConnection() {
+export async function testDatabaseConnection(dbName: string) {
   try {
     // Get the database instance
-    const database = await db.getDb();
+    const database = await db.getDb(dbName);
     // Send a ping to confirm a successful connection
     await database.admin().ping();
     console.log(
